requirements.txt
Django==4.1.7
environs==9.5.0
djangorestframework==3.14.0
django-cors-headers
django-filter==22.1
drf-yasg